+++ src/components/pages/admin/payments.interface.ts
@@ -1,12 +1,9 @@
 import { computed } from 'vue';
 import i18n from '../../../lib/i18n';
 import { type Header } from '../../table/table.interface';
-import {
-  type FilterGroup,
-  type FilterItems,
-  withLabel,
-} from '../../search/search.interface';
+import type { FilterGroup, FilterItems } from '@type';
 import { type PaymentFilterName, paymentFilters } from '@beabee/beabee-common';
+import { withLabel } from '@utils/rules';
 
 const { t } = i18n.global;
