+++ src/components/search/AppSearchRuleOrGroup.vue
@@ -10,15 +10,17 @@
     :rule="rule"
     :readonly="readonly"
     @update:rule="$emit('update:rule', $event)"
+    @remove="$emit('remove')"
   />
 </template>
 <script lang="ts" setup>
 import { isRuleGroup, type Rule, type RuleGroup } from '@beabee/beabee-common';
 import { useI18n } from 'vue-i18n';
 import AppSearchRule from './AppSearchRule.vue';
-import type { FilterGroup, FilterItems } from './search.interface';
 
-defineEmits(['update:rule']);
+import type { FilterGroup, FilterItems } from '@type';
+
+defineEmits(['update:rule', 'remove']);
 defineProps<{
   filterGroups: FilterGroup[];
   filterItems: FilterItems;
