+++ src/components/search/operators/ArgInput.vue
@@ -39,7 +39,7 @@ import AppRadioGroup from '../../forms/AppRadioGroup.vue';
 import AppSelect from '../../forms/AppSelect.vue';
 import DateInput from './DateInput.vue';
 import { useI18n } from 'vue-i18n';
-import type { FilterItem } from '../search.interface';
+import type { FilterItem } from '@type';
 
 const emit = defineEmits(['update:modelValue']);
 const props = defineProps<{
