+++ src/type/filter-group.ts
@@ -1,0 +1,4 @@
+export interface FilterGroup<T extends string = string> {
+  label: string;
+  items: T[];
+}
