+++ src/type/index.ts
@@ -44,6 +44,8 @@ export * from './create-reset-device-data';
 export * from './create-reset-password-data';
 export * from './create-segment-data';
 export * from './delete-contact-mfa-data';
+export * from './filter-group';
+export * from './filter-item';
 export * from './force-update-contribution-data';
 export * from './get-api-key-data';
 export * from './get-api-keys-query';
@@ -86,6 +88,7 @@ export * from './login-data';
 export * from './noop';
 export * from './notice-data';
 export * from './payment-flow-params';
+export * from './rule-group-with-empty';
 export * from './serial';
 export * from './set-contribution-data';
 export * from './set-mfa-steps';
