+++ src/type/rule-group-with-empty.ts
@@ -1,0 +1,6 @@
+import type { Rule, RuleGroup } from '@beabee/beabee-common';
+
+export interface RuleGroupWithEmpty {
+  condition: RuleGroup['condition'];
+  rules: (Rule | RuleGroup | null)[];
+}
